requirements.txt
# Copyright 2017 Canonical Ltd.
# Licensed under the LGPLv3, see LICENCE file for details.

requests==2.18.4
PyNaCl==1.1.2
pymacaroons==0.12.0
six==1.11.0
protobuf==3.4.0
pyRFC3339==1.0
pytz==2017.2'
